fixup: reduce code duplications
