+++ libs/providers/flagd/src/e2e/step-definitions/reconnect.ts
@@ -12,8 +12,17 @@ export const reconnectStepDefinitions: StepDefinitions = ({ given, and, when, th
   let readyRunCount = 0;
   let errorHandlerRun = 0;
 
+  /**
+   * This describe block and retry settings are calibrated to gRPC's retry time
+   * and our testing container's restart cadence.
+   */
+  const retryTimes = 240;
+  jest.retryTimes(retryTimes);
+  const retryDelayMs = 5000;
+
   beforeAll((done) => {
     client.addHandler(ProviderEvents.Ready, () => {
+      readyRunCount++;
       done();
     });
   });
@@ -23,20 +32,17 @@ export const reconnectStepDefinitions: StepDefinitions = ({ given, and, when, th
       errorRunCount++;
     });
 
-    client.addHandler(ProviderEvents.Ready, () => {
-      readyRunCount++;
-    });
   });
   then('the PROVIDER_READY handler must run when the provider connects', async () => {
     // should already be at 1 from `beforeAll`
     expect(readyRunCount).toEqual(1);
   });
   and("the PROVIDER_ERROR handler must run when the provider's connection is lost", async () => {
-    await new Promise((resolve) => setTimeout(resolve, 10000));
+    await new Promise((resolve) => setTimeout(resolve, retryDelayMs));
     expect(errorRunCount).toBeGreaterThan(0);
   });
   and('when the connection is reestablished the PROVIDER_READY handler must run again', async () => {
-    await new Promise((resolve) => setTimeout(resolve, 10000));
+    await new Promise((resolve) => setTimeout(resolve, retryDelayMs));
     expect(readyRunCount).toBeGreaterThan(1);
   });
   when('a flagd provider is set and initialization is awaited', () => {
