+++ libs/shared/flagd-core/src/e2e/index.ts
@@ -1,0 +1,8 @@
+export const E2E_CLIENT_NAME = 'e2e';
+
+export const IMAGE_VERSION = 'v0.5.13';
+
+export function getGherkinTestPath(file: string, modulePath = 'test-harness/gherkin/'): string {
+  // TODO: find a way to resolve this in a generic manner - currently this works, because of the file structure
+  return `<rootdir>/../../../../../shared/flagd-core/${modulePath}${file}`;
+}
